[MOV] sale_management: split sale_order_template.py
