+++ addons/sale_management/models/__init__.py
@@ -7,3 +7,5 @@ from . import sale_order
 from . import sale_order_line
 from . import sale_order_option
 from . import sale_order_template
+from . import sale_order_template_line
+from . import sale_order_template_option
